The Logged In User's Current Directory Is Tracked
When the server starts it now checks if the login file is there, creating it if it isn't so the progam doesn't crash. The server then creates a Users directory for the users folders. The server then creates a folder for each user named the same as the users name and puts them in the Users folder. When the user logs in, there current directory is tracked based on their username.